feat: log the authorization URL
It's very important for folks to understand what that authorization URL does and does not include.

For example, if folks don't include a particular parameter in the authorization URL, then the OAuth flow isn't going to return the information that they want to retrieve.

Logging the authorization URL makes it easier for them to see and debug if they did or did not actually include a parameter for the OAuth flow.