feat: log the Identity Token
The Identity Token is a very useful thing for developers to not only understand, but also to see so they can debug more easily.

Here we log the Identity Token so folks can see which bits of information are contained within the Identity Token, _as they requested_.

This change goes hand-in-hand with the logging of the authorization URL. If folks forget to add a parameter (e.g. a claim) to the authorization URL, then they definitely won't get that value back in the Identity Token.

The hope here is that debugging will be easier if folks can see these things in the logs.