feat: Unique Customer Identifier (commented out)
This provides lines of code--commented out--and docs to explain how to retrieve the Unique Customer Identifier which is a Restricted Claim.

Since this is a Restricted Claim, we don't want folks to get confused by seeing "undefined" in the logs.

The hope here is that the developer can reach out to the administrator at the Financial Institution to get the restricted claim turned on for their External Application. Then, they can uncomment these two lines and see the Unique Customer Identifier in the logs.