Reduce friction points (#22)
* feat: redirect '/' to '/static' route

Folks sometimes click the log statement "App running at http://localhost:8080" which navigates them to the '/' route which wasn't previously handled. This previously resulted in confusion because their web browser would say "Cannot GET /" and they would think that something was broken.

Redirecting folks to the '/static' route to display the static HTML version of the plugin is probably the best choice here to reduce confusion and also give folks the warm fuzzy feeling that their local server is actually functioning correctly.

* feat: log the authorization URL

It's very important for folks to understand what that authorization URL does and does not include.

For example, if folks don't include a particular parameter in the authorization URL, then the OAuth flow isn't going to return the information that they want to retrieve.

Logging the authorization URL makes it easier for them to see and debug if they did or did not actually include a parameter for the OAuth flow.

* feat: log the Identity Token

The Identity Token is a very useful thing for developers to not only understand, but also to see so they can debug more easily.

Here we log the Identity Token so folks can see which bits of information are contained within the Identity Token, _as they requested_.

This change goes hand-in-hand with the logging of the authorization URL. If folks forget to add a parameter (e.g. a claim) to the authorization URL, then they definitely won't get that value back in the Identity Token.

The hope here is that debugging will be easier if folks can see these things in the logs.

* feat: Unique Customer Identifier (commented out)

This provides lines of code--commented out--and docs to explain how to retrieve the Unique Customer Identifier which is a Restricted Claim.

Since this is a Restricted Claim, we don't want folks to get confused by seeing "undefined" in the logs.

The hope here is that the developer can reach out to the administrator at the Financial Institution to get the restricted claim turned on for their External Application. Then, they can uncomment these two lines and see the Unique Customer Identifier in the logs.